Spellcheck fix

+++ utils/check.sh
@@ -1,14 +1,19 @@
 #!/bin/bash
 
 exceptions=`cat exceptions.dict`
+warnings="todo fixme"
 
 cd ../
 
 function print_error {
 	while read -r data; do
-		if [[ "$exceptions" != *$data* ]]; then
-			echo -e "\033[0;31merror:\033[0m ${data}"
-			echo $data >> errors.txt
+		if [[ "$warnings" == *$data* ]]; then
+			echo -e "\033[0;33mwarning:\033[0m ${data}"
+		else
+			if [[ "$exceptions" != *$data* ]]; then
+				echo -e "\033[0;31merror:\033[0m ${data}"
+				echo $data >> errors.txt
+			fi
 		fi
   	done
 }
@@ -22,9 +27,13 @@ find -name "*.md" | while read file_path;
 	do echo -e "\n\033[1;36m${file_path}\033[0m" && check_file $file_path;
 done
 
-errors=$(cat errors.txt | sort | uniq | wc -l)
+errors=0
 
-rm errors.txt
+if [[ -f errors.txt ]]; then
+	errors=$(cat errors.txt | sort | uniq | wc -l)
+	rm errors.txt
+fi
+
 echo -e "\nFound ${errors} unique errors."
 
 if [[ $errors > 0 ]]; then
